factor out common code used for stations and districts
